requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-experimental>=0.0.50

# Database connectivity
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# Streamlit for web interface
streamlit>=1.28.0

# Additional dependencies
pydantic>=2.0.0
sqlalchemy>=2.0.0
requests>=2.25.0
python-dotenv>=1.0.0
